index based finders (swift, quasar) moved to index module
git-svn-id: http://svn.seqan.de/seqan/trunk@5578 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/seqan/index/find_swift.h
@@ -1001,7 +1001,7 @@ inline bool _swiftMultiProcessQGram(
 			}
 			else
 			{
-				if ((*bkt).lastIncrement + bucketParams.tabooLength > curPos) 
+				if ((__int64)(*bkt).lastIncrement + bucketParams.tabooLength > curPos)
 					goto checkOverlap;	// increment only once per sequence			
 				hitCount = (*bkt).counter + 1;
 			}
